changed to u128

+++ src/factorial.rs
@@ -18,7 +18,7 @@
 /// assert_eq!(bens_number_theory::factorial::factorial(5), 120);
 /// assert_eq!(bens_number_theory::factorial::factorial(10), 3628800);
 /// ```
-pub fn factorial(n: u64) -> u64 {
+pub fn factorial(n: u128) -> u128 {
     match n {
         0 | 1 => 1,
         _ => (2..=n).product(),
